Fixed warning about loss of data when converting `size_t' to `int', now we simply use `size_t'.
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@74092 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/C/run-time/except.c
@@ -2434,7 +2434,7 @@ rt_private void extend_trace_string(char *line)
 	 * performed if necessary.
 	 */
 	RT_GET_CONTEXT
-	int l_buf_sz = 0;
+	size_t l_buf_sz = 0;
 
 	if ((ex_string.size - ex_string.used) > strlen(line)) {
 		strcpy (ex_string.area + ex_string.used, line);
